#620: Refactor engine to plugin
