+++ config.yml
@@ -1,5 +1,6 @@
 plugins:
   - lando-core
+  - lando-engine
   # - lando-events
   # - lando-proxy
   # - lando-recipes
