+++ lib/config.js
@@ -57,6 +57,58 @@ exports.merge = function(old, fresh) {
   });
 };
 
+/**
+ * Updates the PATH with dir. This adds dir to the beginning of PATH.
+ *
+ * @since 3.0.0
+ * @param {String} dir - The dir to add
+ * @returns {String} Updated PATH string
+ * @example
+ *
+ * // Update the path
+ * var config.path = config.updatePath(path);
+ */
+exports.updatePath = function(dir) {
+
+  // Determine the path string
+  var p = (process.platform === 'win32') ? 'Path' : 'PATH';
+
+  // Update process.env and return the path
+  if (!_.startsWith(process.env[p], dir)) {
+    process.env[p] = [dir, process.env[p]].join(path.delimiter);
+  }
+
+  // Return
+  return process.env[p];
+
+};
+
+/**
+ * Strips process.env of all envvars with PREFIX
+ *
+ * @since 3.0.0
+ * @param {String} prefix - The prefix to strip
+ * @returns {Object} Updated process.env
+ * @example
+ *
+ * // Reset the process.env without any LANDO_ prefixed envvars
+ * process.env = config.stripEnv('LANDO_');
+ */
+exports.stripEnv = function(prefix) {
+
+  // Strip it down
+  _.each(process.env, function(value, key) {
+    if (_.includes(key, prefix)) {
+      delete process.env[key];
+    }
+  });
+
+  // Return
+  return process.env;
+
+};
+
+
 /*
  * Define default config
  */
@@ -67,8 +119,27 @@ exports.defaults = function() {
   var packageJson = require(path.join(__dirname, '..', 'package.json'));
   var srcRoot = path.resolve(__dirname, '..');
 
-  // Load up our initial env
-  var env = process.env || {};
+  // The default config
+  var config = {
+    configFilename: configFilename,
+    configSources: [path.join(srcRoot, configFilename)],
+    env: process.env,
+    home: os.homedir(),
+    logLevel: 'debug',
+    logLevelConsole: 'warn',
+    node: process.version,
+    os: {
+      type: os.type(),
+      platform: os.platform(),
+      release: os.release(),
+      arch: os.arch()
+    },
+    pluginDirs: [srcRoot],
+    srcRoot: srcRoot,
+    sysConfRoot: getSysConfRoot(),
+    userConfRoot: path.join(os.homedir(), '.lando'),
+    version: packageJson.version
+  };
 
   // Set Path environmental variable if we are on windows so we get access
   // to things like ssh.exe
@@ -87,32 +158,12 @@ exports.defaults = function() {
     //
     // See https://github.com/kalabox/kalabox/issues/342
     _.forEach([gitBin1, gitBin2, gitBin3], function(gBin) {
-      if (fs.existsSync(gBin) && !_.startsWith(env.Path, gBin)) {
-        env.Path = [gBin, env.Path].join(';');
+      if (fs.existsSync(gBin) && !_.startsWith(process.env.Path, gBin)) {
+        process.env.Path = [gBin, process.env.Path].join(';');
       }
     });
-  }
 
-  var config = {
-    configFilename: configFilename,
-    configSources: [path.join(srcRoot, configFilename)],
-    env: env,
-    home: os.homedir(),
-    logLevel: 'debug',
-    logLevelConsole: 'warn',
-    node: process.version,
-    os: {
-      type: os.type(),
-      platform: os.platform(),
-      release: os.release(),
-      arch: os.arch()
-    },
-    pluginDirs: [srcRoot],
-    srcRoot: srcRoot,
-    sysConfRoot: getSysConfRoot(),
-    userConfRoot: path.join(os.homedir(), '.lando'),
-    version: packageJson.version
-  };
+  }
 
   // Return default config
   return config;
