+++ lib/shell.js
@@ -21,7 +21,7 @@
 'use strict';
 
 // Modules
-var _ = require('./node')._;
+var _ = require('lodash')._;
 var _shell = require('shelljs');
 var _spawn = require('child_process').spawn;
 var _esc = require('shell-escape');
@@ -44,13 +44,11 @@ module.exports = function(logger) {
 
     // IF the app has process envars lets merge though in
     if (opts.app && !_.isEmpty(opts.app.processEnv)) {
-      //return _.merge(config.env, opts.app.processEnv);
-      return _.merge({}, opts.app.processEnv);
+      return _.merge(process.env, opts.app.processEnv);
     }
 
-    // Else reutrn the normal config
-    //return config.env;
-    return {};
+    // Else reutrn the normal emv
+    return process.env;
 
   }
