+++ lib/user.js
@@ -30,7 +30,7 @@ var shell = require('sync-exec');
  * // Get the id of the user.
  * var userId = lando.user.getEngineUserId();
  */
-exports.getEngineUserId = function() {
+exports.getUid = function() {
   if (process.platform === 'win32') {
     return 1000;
   }
@@ -57,7 +57,7 @@ exports.getEngineUserId = function() {
  * // Get the id of the user.
  * var groupId = lando.user.getEngineUserGid();
  */
-exports.getEngineUserGid = function() {
+exports.getGid = function() {
   if (process.platform === 'win32') {
     return 50;
   }
