+++ plugins/NEEDTOPUTSOMEWHERE.js
@@ -3,7 +3,6 @@ var APP_ROOT_DIRNAME = process.env.LANDO_CORE_APP_ROOT_DIRNAME || 'Lando';
 var LANDOFILE_NAME = process.env.LANDO_CORE_LANDOFILE_NAME || '.lando.yml';
 var user = require('./user');
 
-
   // App identifier
   var appId = [
     _.get(opts, 'app.name', 'unknown'),
@@ -34,27 +33,10 @@ var user = require('./user');
 //appsRoot: path.join(env.home, APP_ROOT_DIRNAME),
 //appRegistry: path.join(env.userConfRoot, 'appRegistry.json'),
 //cache: true,
-//composeBin: env.composeBin,
-//composeVersion: '3.2',
-//containerGlobalEnv: {},
-//dockerBin: env.dockerBin,
-//dockerBinDir: env.dockerBinDir,
-//engineId: user.getEngineUserId(),
-//engineGid: user.getEngineUserGid(),
 //loadPassphraseProtectedKeys: false,
 
-/*
 
-/**
- * Translate a name for use by docker-compose eg strip `-` and `.` and
- * @TODO: possibly more than that
- *
- * @since 3.0.0
- */
- /*
-exports.dockerComposify = function(data) {
-  return data.replace(/-/g, '').replace(/\./g, '');
-};
+/*
 
 
 /**
@@ -157,71 +139,6 @@ exports.scanUrls = function(urls, opts) {
 
 };
 
-// Add docker executables path to path to handle weird situations where
-// the user may not have machine in their path
-var pathString = (process.platform === 'win32') ? 'Path' : 'PATH';
-var binPath = getDockerBinPath();
-if (!_.startsWith(env[pathString], binPath)) {
-  env[pathString] = [binPath, process.env[pathString]].join(path.delimiter);
-}
-
-/*
- * Helper to get location of docker bin directory
- */
-/*
-var getDockerBinPath = function() {
-  switch (process.platform) {
-    case 'darwin':
-      return path.join('/Applications/Docker.app/Contents/Resources', 'bin');
-    case 'linux':
-      return path.join(getSysConfRoot(), 'bin');
-    case 'win32':
-      var programFiles = process.env.ProgramW6432 || process.env.ProgramFiles;
-      return path.join(programFiles + '\\Docker\\Docker\\resources\\bin');
-  }
-};
-*/
-
-/*
- * Get docker compose binary path
- */
- /*
-var getComposeExecutable = function() {
-
-  // Get compose bin path
-  var composePath = getDockerBinPath();
-  var composeBin = path.join(composePath, 'docker-compose');
-
-  // Return exec based on path
-  switch (process.platform) {
-    case 'darwin': return composeBin;
-    case 'linux': return composeBin;
-    case 'win32': return composeBin + '.exe';
-  }
-
-};
-*/
-
-/*
- * This should only be needed for linux
- */
- /*
-var getDockerExecutable = function() {
-
-  // Get docker bin path
-  var dockerPath = getDockerBinPath();
-  var dockerBin = path.join(dockerPath, 'docker');
-
-  // Return exec based on path
-  switch (process.platform) {
-    case 'darwin': return dockerBin;
-    case 'linux': return '/usr/bin/docker';
-    case 'win32': return dockerBin + '.exe';
-  }
-
-};
-*/
-
 /**
  * This adds a function so apps scan their URLS when they start up
  *
@@ -285,7 +202,6 @@ module.exports = function(lando) {
     lando.init = require('./init')(lando);
 
     // Load our tasks
-    lando.tasks.add('config', require('./tasks/config')(lando));
     lando.tasks.add('destroy', require('./tasks/destroy')(lando));
     lando.tasks.add('info', require('./tasks/info')(lando));
     lando.tasks.add('list', require('./tasks/list')(lando));
@@ -296,7 +212,6 @@ module.exports = function(lando) {
     lando.tasks.add('share', require('./tasks/share')(lando));
     lando.tasks.add('start', require('./tasks/start')(lando));
     lando.tasks.add('stop', require('./tasks/stop')(lando));
-    lando.tasks.add('version', require('./tasks/version')(lando));
 
   });
 
@@ -564,3 +479,4 @@ module.exports = function(lando) {
   });
 
 };
+*/
